refactor(ChessGame): ChessController에서 ChessGame로 클래스명 변경
-게임 진행과 출력 순서를 모두 알고 있어 Controller 역할을 하지 않다는 다고 판단되어 ChessGame으로 클래스명 변경함.
