+++ src/main/java/chessgame/Application.java
@@ -1,10 +1,10 @@
 package chessgame;
 
-import chessgame.controller.ChessController;
+import chessgame.domain.ChessGame;
 
 public class Application {
     public static void main(String[] args) {
-        final var controller = new ChessController();
+        final var controller = new ChessGame();
 
         controller.start();
     }
