+++ src/main/java/chessgame/domain/ExceptionHandler.java
@@ -1,4 +1,4 @@
-package chessgame.controller;
+package chessgame.domain;
 
 import java.util.function.Supplier;
 import chessgame.view.OutputView;
